chore: Move loadenv to /lib (#8219)
