+++ src/dev.js
@@ -5,7 +5,7 @@ require("@babel/register")({
   extensions: [".ts", ".js", ".tsx", ".jsx"],
   plugins: ["babel-plugin-dynamic-import-node"],
 })
-require("./loadenv")
+require("./lib/loadenv")
 
 const express = require("express")
 const path = require("path")
