+++ src/index.js
@@ -3,7 +3,7 @@
 // This must come before any other instrumented module.
 // See https://docs.datadoghq.com/tracing/languages/nodejs/ for more info.
 import "./lib/datadog"
-import "./loadenv"
+import "./lib/loadenv"
 
 // Needs to be first, due to sharify side-effects.
 import { initializeMiddleware } from "./middleware"
